+++ cms/src/main/java/com/zrzhen/zetty/cms/service/ArticleService.java
@@ -1,6 +1,5 @@
 package com.zrzhen.zetty.cms.service;
 
-import java.math.BigInteger;
 import java.util.Map;
 
 /**
@@ -10,13 +9,14 @@ public class ArticleService {
 
     /**
      * 判断访问者是否为文章作者
+     *
      * @param result
      * @return
      */
     public static boolean permission(Map<String, Object> result) {
         if (result.get("status").equals(1)) {
             Long userid = UserService.getUserid();
-            Long userId2 = ((BigInteger) result.get("userid")).longValue();
+            Long userId2 = (Long) result.get("userid");
             if (userid == null || !userid.equals(userId2)) {
                 return false;
             }
